Add show command and refactor init
Add a `show` command that displays details about a transform not shown
in the `list` command.

Add an `IJetfireSubCommand` interface, and take a list of these in the
`JetfireRootCommand` class instead of an explicit list of all the
commands. Should make extending further easier as there are fewer
places to add the new command.